+++ views/js/listViewRunner.js
@@ -27,7 +27,7 @@ source.onmessage = async function(e) {
     } else {
       printerInfo = res.printerInfo;
       if (res.clientSettings.listView.currentOp) {
-        currentOperations(res.currentOperations, res.currentOperationsCount);
+        currentOperations(res.currentOperations, res.currentOperationsCount, res.printerInfo);
       }
       updateState(res.printerInfo, res.clientSettings.listView);
     }