+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
  - Show printer index on printer with no name
  - Current Operations Card on view's now displays the predicted end date and time, calculated from the current date + print time remaining.
  - More Checks into the API/Websocket connection... if it fails to grab your API Key it will now warn you. You will see "No-API" in the state. Pressing the refresh button for your printer on dashboard will attempt a reconnect. 
+ - Current Operations allows you to now de-select a file and remove it from the view. When a print is complete, you will see a "Harvest Your Print!" button, click it to remove the printer from that View.
 
 ### Changed
  - Dashboard layout has been updated with a fixed sidebar. Medium sized screens and above will display this. 
@@ -23,7 +24,8 @@ All notable changes to this project will be documented in this file.
  - async information reception from server -> client. 
  - State should now show if Serial Error occurs on OctoPrint.
  - Shutdown printers, that have been scanned by the farm will now wait 15-seconds before trying to grab an API connection. This should help those with websockets not connecting initially.
-
+ - History now captures and displays selected rolls of filament. No usage available from OctoPrint currently. 
+ 
 ### Removed
 
 ### Deprecated
@@ -38,6 +40,8 @@ All notable changes to this project will be documented in this file.
  - Camera View Current Operations now updates with printers.
  - Fixed issue with Current Operations not clearing final finished or ended print card. 
  - Issue #22 - File Manager crashes when using Canvas Hub.
+ - Docker should now correctly put logs into log folder.
+ - Freshly created filament is now found and saved in the history logs. 
 
 ### Security